Change type of convert param to unsigned long long

+++ converter.c
@@ -7,7 +7,7 @@
  * @lowercase: flag if hexa values need to be lowercase
  * Return: result string
  */
-char *convert(long int num, int base, int lowercase)
+char *convert(unsigned long long num, int base, int lowercase)
 {
 	static char *rep;
 	static char buffer[50];
